+++ distant-ssh2/src/api.rs
@@ -1,4 +1,4 @@
-use std::collections::{HashMap, HashSet};
+use std::collections::{HashMap, HashSet, VecDeque};
 use std::io;
 use std::path::PathBuf;
 use std::sync::{Arc, Weak};
@@ -9,13 +9,15 @@ use async_once_cell::OnceCell;
 use async_trait::async_trait;
 use distant_core::net::server::ConnectionCtx;
 use distant_core::protocol::{
-    Capabilities, CapabilityKind, DirEntry, Environment, FileType, Metadata, ProcessId, PtySize,
-    SystemInfo, UnixMetadata,
+    Capabilities, CapabilityKind, DirEntry, Environment, FileType, Metadata, Permissions,
+    ProcessId, PtySize, SetPermissionsOptions, SystemInfo, UnixMetadata,
 };
 use distant_core::{DistantApi, DistantCtx};
 use log::*;
 use tokio::sync::{mpsc, RwLock};
-use wezterm_ssh::{FilePermissions, OpenFileType, OpenOptions, Session as WezSession, WriteMode};
+use wezterm_ssh::{
+    FilePermissions, OpenFileType, OpenOptions, Session as WezSession, Utf8PathBuf, WriteMode,
+};
 
 use crate::process::{spawn_pty, spawn_simple, SpawnResult};
 use crate::utils::{self, to_other_error};
@@ -86,6 +88,9 @@ impl DistantApi for SshDistantApi {
         // TODO: Could we have external search using ripgrep's JSON lines API?
         capabilities.take(CapabilityKind::Search);
         capabilities.take(CapabilityKind::CancelSearch);
+
+        // Broken via wezterm-ssh, so not supported right now
+        capabilities.take(CapabilityKind::SetPermissions);
 
         Ok(capabilities)
     }
@@ -664,7 +669,7 @@ impl DistantApi for SshDistantApi {
             // Check that owner, group, or other has write permission (if not, then readonly)
             readonly: metadata
                 .permissions
-                .map(FilePermissions::is_readonly)
+                .map(|x| !x.owner_write && !x.group_write && !x.other_write)
                 .unwrap_or(true),
             accessed: metadata.accessed.map(u128::from),
             modified: metadata.modified.map(u128::from),
@@ -682,6 +687,136 @@ impl DistantApi for SshDistantApi {
             }),
             windows: None,
         })
+    }
+
+    #[allow(unreachable_code)]
+    async fn set_permissions(
+        &self,
+        ctx: DistantCtx<Self::LocalData>,
+        path: PathBuf,
+        permissions: Permissions,
+        options: SetPermissionsOptions,
+    ) -> io::Result<()> {
+        debug!(
+            "[Conn {}] Setting permissions for {:?} {{permissions: {:?}, options: {:?}}}",
+            ctx.connection_id, path, permissions, options
+        );
+
+        // Unsupported until issue resolved: https://github.com/wez/wezterm/issues/3784
+        return Err(io::Error::new(
+            io::ErrorKind::Unsupported,
+            "Unsupported until issue resolved: https://github.com/wez/wezterm/issues/3784",
+        ));
+
+        let sftp = self.session.sftp();
+
+        macro_rules! set_permissions {
+            ($path:ident, $metadata:ident) => {{
+                let mut current = Permissions::from_unix_mode(
+                    $metadata
+                        .permissions
+                        .ok_or_else(|| to_other_error("Unable to read file permissions"))?
+                        .to_unix_mode(),
+                );
+
+                current.apply_from(&permissions);
+
+                $metadata.permissions =
+                    Some(FilePermissions::from_unix_mode(current.to_unix_mode()));
+
+                println!("set_metadata for {:?}", $path.as_path());
+                sftp.set_metadata($path.as_path(), $metadata)
+                    .compat()
+                    .await
+                    .map_err(to_other_error)?;
+
+                if $metadata.is_dir() {
+                    Some($path)
+                } else {
+                    None
+                }
+            }};
+            ($path:ident) => {{
+                let mut path = Utf8PathBuf::try_from($path).map_err(to_other_error)?;
+
+                // Query metadata to determine if we are working with a symlink
+                println!("symlink_metadata for {:?}", path);
+                let mut metadata = sftp
+                    .symlink_metadata(&path)
+                    .compat()
+                    .await
+                    .map_err(to_other_error)?;
+
+                // If we are excluding symlinks and this is a symlink, then we're done
+                if options.exclude_symlinks && metadata.is_symlink() {
+                    None
+                } else {
+                    // If we are following symlinks and this is a symlink, then get the real path
+                    // and destination metadata
+                    if options.follow_symlinks && metadata.is_symlink() {
+                        println!("read_link for {:?}", path);
+                        path = sftp
+                            .read_link(path)
+                            .compat()
+                            .await
+                            .map_err(to_other_error)?;
+
+                        println!("metadata for {:?}", path);
+                        metadata = sftp
+                            .metadata(&path)
+                            .compat()
+                            .await
+                            .map_err(to_other_error)?;
+                    }
+
+                    set_permissions!(path, metadata)
+                }
+            }};
+        }
+
+        let mut paths = VecDeque::new();
+
+        // Queue up our path if it is a directory
+        if let Some(path) = set_permissions!(path) {
+            paths.push_back(path);
+        }
+
+        if options.recursive {
+            while let Some(path) = paths.pop_front() {
+                println!("read_dir for {:?}", path);
+                let paths_and_metadata =
+                    sftp.read_dir(path).compat().await.map_err(to_other_error)?;
+                for (mut path, mut metadata) in paths_and_metadata {
+                    if options.exclude_symlinks && metadata.is_symlink() {
+                        println!("skipping symlink for {:?}", path);
+                        continue;
+                    }
+
+                    // If we are following symlinks, then adjust our path and metadata
+                    if options.follow_symlinks && metadata.is_symlink() {
+                        println!("read_link for {:?}", path);
+                        path = sftp
+                            .read_link(path)
+                            .compat()
+                            .await
+                            .map_err(to_other_error)?;
+
+                        println!("metadata for {:?}", path);
+                        metadata = sftp
+                            .metadata(&path)
+                            .compat()
+                            .await
+                            .map_err(to_other_error)?;
+                    }
+
+                    if let Some(path) = set_permissions!(path, metadata) {
+                        paths.push_back(path);
+                    }
+                }
+            }
+        }
+
+        Ok(())
     }
 
     async fn proc_spawn(